Remove unnecesary postfix restart.
The Problem:
============
Having this service resource like this causes postfix to be restarted on
every chef run, which will make it unable to serve requests and cause
performance problems as workers spin back up.

As the postfix cookbook notifies the postfix service to `reload` on any
configuration change, it's not required here.

The Solution:
=============
Remove the service resource and stop postfix from restarting every half
an hour.