Add proper support for opendkim v2.11+ by updating config file
Support was previously provided by reverting the service file to an old
version to match the old config, however the proper fix would have been
updating the service to match the new config.

`autorestart` and `sender_headers` attributes were removed as there's no
reference to them in the new config file.

If the service file looks old (likely had our previous "fix" applied),
we'll update the service file to what the package likely would have
installed to ensure compatibility with the new config file.

This likely now requires Ubuntu 18.04 or newer as the new config file
may not be compatible with versions of opendkim shipped with earlier
versions of Ubuntu.

The new path for keyfiles seems to be /etc/dkimkeys/, however I left the
cookbook default of /etc/mail/ so we don't have to deal with potentially
moving keys for instances that already have them in the cookbook
default.